pyproject.toml
[project]
name = "fastmcp-content-factory"
version = "0.1.0"
description = "FastMCP Multi-Agent Intelligent Content Production System"
readme = "README.md"
authors = [
    { name = "Content Factory Team", email = "team@contentfactory.ai" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "asyncio>=3.4.3",
    "requests>=2.31.0",
    "openai>=1.68.2",
    "tavily-python>=0.7.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "rich>=13.7.1",
    "typer>=0.12.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/content_factory"]

[project.scripts]
content-factory = "content_factory.cli:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
